backend/requirements.txt

django
djangorestframework
django-cors-headers
gunicorn


psycopg2-binary


celery
redis

pdf2image
unidecode
mineru 

python-dotenv
Pygments
torch==2.3.0
torchvision
doclayout-yolo
ultralytics
transformers
ftfy
shapely
pyclipper
omegaconf